+++ src/maite/protocols/providers.py
@@ -2,14 +2,7 @@
 # Subject to FAR 52.227-11 – Patent Rights – Ownership by the Contractor (May 2014).
 # SPDX-License-Identifier: MIT
 
-from maite._internals.protocols.providers import (
-    DatasetProvider,
-    MetricProvider,
-    ModelProvider,
-)
-
-__all__ = [
-    "ModelProvider",
-    "DatasetProvider",
-    "MetricProvider",
-]
+# ----
+# Not exposing provider protocols publicly yet on the grounds that writing provider
+# implementers will likely affect protocol design.
+# ----